coded and validated buffer.js and buffer.test.js
