+++ buffer.js
@@ -1,0 +1,9 @@
+const fs = require('fs');
+
+function getFile(bmp, cb) {
+  fs.readFile(bmp, (err, buffer) => {
+    cb(buffer);
+
+  });
+}
+module.exports = getFile;
